+++ server/auth/auth.go
@@ -587,18 +587,6 @@ func (auth *AuthManager) RequestWithTimeout(user account.AccountID, msg *msgjson
 	return auth.request(user, msg, f, expireTimeout, expire)
 }
 
-// SwapAmounts breaks down the quantities of completed swaps in four rough
-// categories: successfully swapped (Swapped), failed with counterparty funds
-// locked for the long/maker lock time (StuckLong), failed with counterparty
-// funds locked for the short/taker lock time (StuckShort), and failed to
-// initiate swap following match with no funds locked in contracts (Spoofed).
-type SwapAmounts struct {
-	Swapped    int64
-	StuckLong  int64
-	StuckShort int64
-	Spoofed    int64
-}
-
 // userSwapAmountHistory retrieves the summary of recent swap amounts for the
 // given user and market. The user should be connected.
 func (auth *AuthManager) userSwapAmountHistory(user account.AccountID, base, quote uint32) *SwapAmounts {
